pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logo-scraper"
version = "0.1.0"
description = "Fetch and download company logos from websites, logo.dev, and LinkedIn."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
]

[project.scripts]
logo-scraper = "logo_scraper.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
